refactor(generators): Refactor auth generators
