+++ generators/auth_better_auth/generator.go
@@ -3,6 +3,7 @@ package authbetterauth
 import (
 	"embed"
 	"fmt"
+	"strings"
 
 	"github.com/version14/dot/internal/render"
 	"github.com/version14/dot/internal/state"
@@ -28,7 +29,11 @@ func (g *Generator) Generate(ctx *dotapi.Context) error {
 	if err := ctx.State.UpdateJSON("package.json", func(d *state.JSONDoc) error {
 		d.Merge(map[string]interface{}{
 			"dependencies": map[string]interface{}{
-				"better-auth": "^1.2.0",
+				"better-auth":   "^1.2.0",
+				"cookie-parser": "^1.4.7",
+			},
+			"devDependencies": map[string]interface{}{
+				"@types/cookie-parser": "^1.4.8",
 			},
 		})
 		return nil
@@ -43,6 +48,16 @@ func (g *Generator) Generate(ctx *dotapi.Context) error {
 	}
 	updated := existing + fmt.Sprintf("\n# Auth (BetterAuth)\nBETTER_AUTH_SECRET=%s\nBETTER_AUTH_URL=http://localhost:${PORT:-3000}\n", generateSecretPlaceholder())
 	ctx.State.WriteFile(".env.example", []byte(updated), state.ContentRaw)
+
+	// Inject cookie-parser middleware into app.ts
+	if f, ok := ctx.State.GetFile("src/app.ts"); ok {
+		content := string(f.Content)
+		if !strings.Contains(content, "cookieParser") {
+			content = "import cookieParser from 'cookie-parser';\n" + content
+			content = strings.Replace(content, "app.use(express.json());", "app.use(express.json());\napp.use(cookieParser());", 1)
+			ctx.State.WriteFile("src/app.ts", []byte(content), state.ContentRaw)
+		}
+	}
 
 	return nil
 }
