+++ generators/auth_jwt_clean_arch_module/files/src/modules/auth/domain/entities/user.entity.ts
@@ -1,0 +1,7 @@
+export interface UserEntity {
+  id: string;
+  email: string;
+  passwordHash: string;
+  createdAt: Date;
+  updatedAt: Date;
+}
