+++ generators/auth_jwt_clean_arch_module/files/src/modules/auth/domain/interfaces/refresh-token.repository.interface.ts
@@ -1,0 +1,5 @@
+export interface IRefreshTokenRepository {
+  create(data: { token: string; userId: string; expiresAt: Date }): Promise<void>;
+  findByToken(token: string): Promise<{ id: string; userId: string } | null>;
+  deleteById(id: string): Promise<void>;
+}
