+++ generators/auth_jwt_clean_arch_module/files/src/modules/auth/domain/interfaces/user.repository.interface.ts
@@ -1,0 +1,6 @@
+import type { UserEntity } from '../entities/user.entity';
+
+export interface IUserRepository {
+  findByEmail(email: string): Promise<UserEntity | null>;
+  create(data: { email: string; passwordHash: string }): Promise<UserEntity>;
+}
