+++ generators/auth_jwt_mvc_route/files/src/__tests__/auth.db.test.ts.tmpl
@@ -1,0 +1,67 @@
+import request from 'supertest';
+import { beforeEach, describe, expect, it } from 'vitest';
+import app from '../app';
+import { db } from '../db';
+import { refreshTokens, users } from '../db/schema';
+
+beforeEach(async () => {
+  await db.delete(refreshTokens);
+  await db.delete(users);
+});
+
+describe('auth routes (DB)', () => {
+  const email = 'db-test@example.com';
+  const password = 'password123';
+
+  it('POST /auth/register creates user and sets cookies', async () => {
+    const res = await request(app).post('/auth/register').send({ email, password });
+    expect(res.status).toBe(201);
+    const cookies = res.headers['set-cookie'];
+    const cookieArray = Array.isArray(cookies) ? cookies : cookies ? [cookies] : [];
+    expect(cookieArray.some((c: string) => c.includes('access_token'))).toBe(true);
+    expect(cookieArray.some((c: string) => c.includes('refresh_token'))).toBe(true);
+  });
+
+  it('POST /auth/register with duplicate email returns 409', async () => {
+    await request(app).post('/auth/register').send({ email, password });
+    const res = await request(app).post('/auth/register').send({ email, password });
+    expect(res.status).toBe(409);
+  });
+
+  it('POST /auth/login sets cookies', async () => {
+    await request(app).post('/auth/register').send({ email, password });
+    const res = await request(app).post('/auth/login').send({ email, password });
+    expect(res.status).toBe(200);
+    const cookies = res.headers['set-cookie'];
+    const cookieArray = Array.isArray(cookies) ? cookies : cookies ? [cookies] : [];
+    expect(cookieArray.some((c: string) => c.includes('access_token'))).toBe(true);
+  });
+
+  it('GET /auth/me with valid cookie returns user', async () => {
+    const regRes = await request(app).post('/auth/register').send({ email, password });
+    const cookies = regRes.headers['set-cookie'] || [];
+    const res = await request(app).get('/auth/me').set('Cookie', cookies as string[]);
+    expect(res.status).toBe(200);
+    expect(res.body.user).toBeDefined();
+  });
+
+  it('POST /auth/refresh rotates tokens in cookies', async () => {
+    const regRes = await request(app).post('/auth/register').send({ email, password });
+    const cookies = regRes.headers['set-cookie'] || [];
+    const res = await request(app).post('/auth/refresh').set('Cookie', cookies as string[]);
+    expect(res.status).toBe(200);
+    const newCookies = res.headers['set-cookie'];
+    const newCookieArray = Array.isArray(newCookies) ? newCookies : newCookies ? [newCookies] : [];
+    expect(newCookieArray.some((c: string) => c.includes('access_token'))).toBe(true);
+  });
+
+  it('POST /auth/logout clears cookies', async () => {
+    const regRes = await request(app).post('/auth/register').send({ email, password });
+    const cookies = regRes.headers['set-cookie'] || [];
+    const res = await request(app).post('/auth/logout').set('Cookie', cookies as string[]);
+    expect(res.status).toBe(204);
+    const finalCookies = res.headers['set-cookie'];
+    const finalCookieArray = Array.isArray(finalCookies) ? finalCookies : finalCookies ? [finalCookies] : [];
+    expect(finalCookieArray.some((c: string) => c.includes('access_token=;'))).toBe(true);
+  });
+});
