+++ generators/auth_jwt_mvc_route/files/src/controllers/auth.controller.ts.tmpl
@@ -1,0 +1,140 @@
+{{ if .HasDB -}}
+import type { Request, Response } from 'express';
+import type { AuthRequest } from '../shared/middlewares/auth.middleware';
+import bcrypt from 'bcryptjs';
+import { eq } from 'drizzle-orm';
+import { db } from '../db';
+import { refreshTokens, users } from '../db/schema';
+import { signRefreshToken, signToken, verifyToken } from '../shared/services/jwt';
+
+const COOKIE_OPTIONS = {
+  httpOnly: true,
+  secure: process.env.NODE_ENV === 'production',
+  sameSite: 'lax' as const,
+};
+
+async function issueTokens(
+  userId: string,
+  email: string,
+): Promise<{ accessToken: string; refreshToken: string }> {
+  const accessToken = signToken({ id: userId, email });
+  const newRefreshToken = signRefreshToken({ id: userId, email });
+  const decoded = verifyToken<{ exp: number }>(newRefreshToken);
+  const expiresAt = new Date(decoded.exp * 1000);
+  await db.insert(refreshTokens).values({ token: newRefreshToken, userId, expiresAt });
+  return { accessToken, refreshToken: newRefreshToken };
+}
+
+function setAuthCookies(res: Response, tokens: { accessToken: string; refreshToken: string }) {
+  res.cookie('access_token', tokens.accessToken, COOKIE_OPTIONS);
+  res.cookie('refresh_token', tokens.refreshToken, COOKIE_OPTIONS);
+}
+
+export async function register(req: Request, res: Response): Promise<void> {
+  const { email, password } = req.body as { email: string; password: string };
+  const existing = await db.select().from(users).where(eq(users.email, email)).limit(1);
+  if (existing.length > 0) {
+    res.status(409).json({ error: 'Email already in use' });
+    return;
+  }
+  const passwordHash = await bcrypt.hash(password, 10);
+  const result = await db.insert(users).values({ email, passwordHash }).returning();
+  const newUser = result[0];
+  if (!newUser) {
+    res.status(500).json({ error: 'Failed to create user' });
+    return;
+  }
+  const tokens = await issueTokens(newUser.id, newUser.email);
+  setAuthCookies(res, tokens);
+  res.status(201).json({ message: 'Registered successfully' });
+}
+
+export async function login(req: Request, res: Response): Promise<void> {
+  const { email, password } = req.body as { email: string; password: string };
+  const result = await db.select().from(users).where(eq(users.email, email)).limit(1);
+  const user = result[0];
+  if (!user) {
+    res.status(401).json({ error: 'Invalid credentials' });
+    return;
+  }
+  const valid = await bcrypt.compare(password, user.passwordHash);
+  if (!valid) {
+    res.status(401).json({ error: 'Invalid credentials' });
+    return;
+  }
+  const tokens = await issueTokens(user.id, user.email);
+  setAuthCookies(res, tokens);
+  res.json({ message: 'Logged in successfully' });
+}
+
+export async function refresh(req: Request, res: Response): Promise<void> {
+  const refreshToken = req.cookies?.refresh_token;
+  if (!refreshToken) {
+    res.status(400).json({ error: 'Refresh token required' });
+    return;
+  }
+  const getPayload = (): { id: string; email: string; exp: number } | null => {
+    try {
+      return verifyToken<{ id: string; email: string; exp: number }>(refreshToken);
+    } catch {
+      return null;
+    }
+  };
+  const payload = getPayload();
+  if (!payload) {
+    res.status(401).json({ error: 'Invalid refresh token' });
+    return;
+  }
+  const stored = await db
+    .select()
+    .from(refreshTokens)
+    .where(eq(refreshTokens.token, refreshToken))
+    .limit(1);
+  const storedToken = stored[0];
+  if (!storedToken) {
+    res.status(401).json({ error: 'Refresh token not found' });
+    return;
+  }
+  await db.delete(refreshTokens).where(eq(refreshTokens.id, storedToken.id));
+  const tokens = await issueTokens(payload.id, payload.email);
+  setAuthCookies(res, tokens);
+  res.json({ message: 'Tokens refreshed' });
+}
+
+export async function logout(req: Request, res: Response): Promise<void> {
+  const refreshToken = req.cookies?.refresh_token;
+  if (refreshToken) {
+    await db.delete(refreshTokens).where(eq(refreshTokens.token, refreshToken));
+  }
+  res.clearCookie('access_token');
+  res.clearCookie('refresh_token');
+  res.status(204).send();
+}
+
+export async function me(req: AuthRequest, res: Response): Promise<void> {
+  res.json({ user: req.user });
+}
+{{ else -}}
+import type { Request, Response } from 'express';
+import type { AuthRequest } from '../shared/middlewares/auth.middleware';
+
+export async function register(_req: Request, res: Response): Promise<void> {
+  res.status(501).json({ error: 'not implemented' });
+}
+
+export async function login(_req: Request, res: Response): Promise<void> {
+  res.status(501).json({ error: 'not implemented' });
+}
+
+export async function refresh(_req: Request, res: Response): Promise<void> {
+  res.status(501).json({ error: 'not implemented' });
+}
+
+export async function logout(_req: Request, res: Response): Promise<void> {
+  res.status(501).json({ error: 'not implemented' });
+}
+
+export async function me(req: AuthRequest, res: Response): Promise<void> {
+  res.json({ user: req.user });
+}
+{{ end -}}
