+++ generators/auth_jwt_users_schema/manifest.go
@@ -10,6 +10,13 @@ var Manifest = dotapi.Manifest{
 	Outputs: []string{
 		"src/db/schema/users.table.ts",
 	},
+	TestCommands: []dotapi.Command{
+		{Cmd: "test -f .env || cp .env.example .env"},
+		{Cmd: "docker compose down -v 2>/dev/null || true"},
+		{Cmd: "docker compose up -d && sleep 5"},
+		{Cmd: "pnpm exec drizzle-kit push --force"},
+		{Cmd: "bash -c 'pnpm exec vitest run db; EXIT_CODE=$?; docker compose down -v; exit $EXIT_CODE'"},
+	},
 	Validators: []dotapi.Validator{
 		{
 			Name: "auth_jwt_users_schema",
